Refactor category and comment schemas to use ObjectId references; update services and modules for improved data handling and relationships

+++ 10.nest-mongodb-mongoose/src/categories/categories.service.ts
@@ -8,11 +8,13 @@ import { UpdateCategoryDto } from './dto/update-category.dto';
 import { Category } from './schemas/categories.schema';
 import { Model } from 'mongoose';
 import { InjectModel } from '@nestjs/mongoose';
+import { Post } from 'src/posts/schemas/posts.schema';
 
 @Injectable()
 export class CategoriesService {
   constructor(
     @InjectModel(Category.name) private categoryModel: Model<Category>,
+    @InjectModel(Post.name) private postModel: Model<Post>,
   ) {}
 
   async create(createCategoryDto: CreateCategoryDto): Promise<Category> {
@@ -25,7 +27,7 @@ export class CategoriesService {
     return await query.limit(Math.max(1, limit)).skip(Math.max(0, skip)).exec();
   }
 
-  async findOne(id: string): Promise<Category | NotFoundException> {
+  async findOne(id: string): Promise<Category> {
     const category = await this.categoryModel
       .findById(id)
       .populate('posts')
@@ -47,7 +49,15 @@ export class CategoriesService {
   }
 
   async remove(id: string) {
-    await this.findOne(id);
+    const category = await this.findOne(id);
+    if (!category) {
+      throw new NotFoundException(`Category with id ${id} not found`);
+    }
+    // remove the category ID from the associated post's category array
+    await this.postModel.updateOne(
+      { _id: category.posts },
+      { $pull: { category: category } },
+    );
     return await this.categoryModel.findByIdAndDelete(id).exec();
   }
 }
