+++ 10.nest-mongodb-mongoose/src/comments/comments.service.ts
@@ -4,11 +4,13 @@ import { UpdateCommentDto } from './dto/update-comment.dto';
 import { Comment } from './schemas/comments.schema';
 import { Model } from 'mongoose';
 import { InjectModel } from '@nestjs/mongoose';
+import { Post } from 'src/posts/schemas/posts.schema';
 
 @Injectable()
 export class CommentsService {
   constructor(
     @InjectModel(Comment.name) private commentModel: Model<Comment>,
+    @InjectModel(Post.name) private postModel: Model<Post>,
   ) {}
 
   async create(createCommentDto: CreateCommentDto): Promise<Comment> {
@@ -24,7 +26,7 @@ export class CommentsService {
     return await query.limit(Math.max(1, limit)).skip(Math.max(0, skip)).exec();
   }
 
-  async findOne(id: string): Promise<Comment | NotFoundException> {
+  async findOne(id: string): Promise<Comment> {
     const comment = await this.commentModel
       .findById(id)
       .populate('author')
@@ -46,8 +48,17 @@ export class CommentsService {
       .exec();
   }
 
-  async remove(id: string) {
-    await this.findOne(id);
+  async remove(id: string): Promise<Comment | null> {
+    const comment = await this.findOne(id);
+    if (!comment) {
+      throw new NotFoundException(`Comment with id ${id} not found`);
+    }
+    // Remove the comment ID from the associated post's comments array.
+    await this.postModel.updateOne(
+      { _id: comment.post },
+      { $pull: { comments: comment } },
+    );
+    // Delete and return the comment.
     return await this.commentModel.findByIdAndDelete(id).exec();
   }
 }
